+++ src/tests/vtkh/t_vtk-h_clip.cpp
@@ -13,7 +13,7 @@
 #include <vtkh/rendering/PointRenderer.hpp>
 #include <vtkh/rendering/Scene.hpp>
 
-#include "t_test_utils.hpp"
+#include "t_vtkm_test_utils.hpp"
 
 #include <iostream>
