+++ src/tests/vtkh/t_vtk-h_dataset_par.cpp
@@ -10,7 +10,7 @@
 #include <vtkh/vtkh.hpp>
 #include <vtkh/DataSet.hpp>
 #include <vtkm/cont/DataSetBuilderExplicit.h>
-#include "t_test_utils.hpp"
+#include "t_vtkm_test_utils.hpp"
 
 #include <iostream>
 #include <mpi.h>
