+++ src/tests/vtkh/t_vtk-h_gradient.cpp
@@ -12,7 +12,7 @@
 #include <vtkh/filters/VectorMagnitude.hpp>
 #include <vtkh/rendering/RayTracer.hpp>
 #include <vtkh/rendering/Scene.hpp>
-#include "t_test_utils.hpp"
+#include "t_vtkm_test_utils.hpp"
 
 #include <iostream>
