+++ src/tests/vtkh/t_vtk-h_point_renderer.cpp
@@ -11,7 +11,7 @@
 #include <vtkh/rendering/PointRenderer.hpp>
 #include <vtkh/filters/MarchingCubes.hpp>
 #include <vtkh/rendering/Scene.hpp>
-#include "t_test_utils.hpp"
+#include "t_vtkm_test_utils.hpp"
 
 #include <iostream>
