+++ src/tests/vtkh/t_vtk-h_raytracer.cpp
@@ -10,7 +10,7 @@
 #include <vtkh/DataSet.hpp>
 #include <vtkh/rendering/RayTracer.hpp>
 #include <vtkh/rendering/Scene.hpp>
-#include "t_test_utils.hpp"
+#include "t_vtkm_test_utils.hpp"
 
 #include <iostream>
