+++ src/tests/vtkh/t_vtk-h_slice.cpp
@@ -11,7 +11,7 @@
 #include <vtkh/filters/Slice.hpp>
 #include <vtkh/rendering/RayTracer.hpp>
 #include <vtkh/rendering/Scene.hpp>
-#include "t_test_utils.hpp"
+#include "t_vtkm_test_utils.hpp"
 
 #include <iostream>
